Add UpdatesPerTick slider
Adds a slider which allows for changing the number of updates that happen per
render cycle. Current max is 100 because it takes longer than 40 ms to update
more.

Also changes the Pixel scale selection to be a slider.

Changes how the toRender queue works so that it can build up multiple changes
from sequential updates, and then render them all at once. This also required
there to be a function that clears out the queue after the update is performed.

Lastly, adds helper methods to Board and Tile to use coords instead of tiles.