Fix trails and add debugging tools
Trails were not decaying when a second trail was added to a tile. The second
trail was not being initialized to 0 so it would become NaN when adding the
strength to undefined. While debugging I refactored updateTrails to use a Set
of tiles instead of an Array of coords.

This commit also adds two debugging tools. The first is logging the amount of
time each update/render takes. It is currently commented out, but this could
easily be gated around a checkbox. Also added is the ability to click on any
tile to 'watch' it and have a JSON hash printed to the bottom of the screen.