Refactor update loop
Refactors the update loop to drive itself with setTimeouts rather than
setInterval. This way it will never get ahead of itself if the loop execution
time becomes longer than the update interval.

Also adds a Pause/Run button and a Step button.

Lastly, reorders the UI with the controls on top, and scores down below.