Improve ant spiraling logic
Makes a more general spiraling solution that is dependent on the location of
the ant in relation to the spiral point rather than a rigid formula. The spiral
function takes a delta param which determines how rapidly it should spiral out.

Adjusts a few other params, such as when to move randomly, and how far the
queen should travel before establishing a new colony.