Add editor and some CSS
Adds the editor the left pane, and moves the gameArea + controls to the top right pane.

Moves player info and tile info to the bottom right pane.

Sliders and buttons are split into two div groups so they can be stacked on top of one another.

Adds some new styling for player scores.