Add wheel zooming and mouse panning
Adds some basic zooming in and out with the mouse wheel. Also adds some basic
panning by clicking and dragging.  While playing around with panning on zoomed
in screens I discovered that centering in a flexbox causes the overflow to not
calculate properly, so you lose the left and top portion of the map. This will
require another commit to fix.