Refactor AceEditor into EditorPane
Creates a new EditorPane component to handle all editor related things,
along with the editor buttons/controls.

Also styles the editor to take up all the space the buttons do not, putting
the buttons on the bottom.

Lastly, scrolls the editor to the top of the page and focuses it on load.