Add ability to edit a local file and watch for changes
Adds an eye icon button to the EditiorControls that when clicked allows the
user to select a file. The code is loaded into the editor, and the icon becomes
an eye with a slash through it.  The selected file is then polled for updates,
and any changes are pulled into the editor, and the game is started again.
Pressing the eye with a slash through it stops watching the file for changes,
but the last pulled in code remains.