Fix Hive Console styling and add an env
Fixes the styling issue on Hive Console by adding proper overflow tags to the
right split Pane and HiveConsole.  Also uses flexbox to determine height of the
tabs and tab-content.

Fixes a bug where after printing to the console it would only scroll to the
bottom after new input. When the component updates we call scrollToBottom().
This causes a small flicker, but is acceptable for now.

Lastly, adds a fake environment to the console by keeping track of succesful
evals, and then joining all the old text lines to the new one with ';'.