Adds text editor console logging
Because the react Console component requires a call to its handler to
properly log a message, the message had to be set in the component's
state.

Currently, if the text editor sends an object to the console for
logging, React throws an error saying "Objects are not valid as a React
child...". Another commit will be required to fix this.