Changes and adds to player info pane
This commit adds additional information to the player info pane such as
number of ants killed/died etc. The info is now split into tabs for each
player and is displayed in tables.

Variables for tracking this info were added to the Player object.
The player object stored in the Hive component was also updated to help
generate the tables.

To do: styling...