Adds Total Ants line graph to PlayerInfo
This commit adds react-easy-chart to package.json for rendering line
graphs. 4 sets of data are stored on the hive component for rendering
TotalAnts, TotalFood, AntsKilled and DeadAnts line graphs.

Todo, add handlers to switch which set of data are being watched.
Figure out how to size the graph appropriately in the split pane.