AIManagerModal: Add autoFocus and select to name
When adding a new AI we want to auto focus and select the name. Since this will
always be the last one we can put auto focus on all the inputs, and the last
one will be the one selected. Since we don't want to focus when we first enter
the modal we have to manage an addedAI bool in state, and clear this when we
leave the modal.